==> 01_load_isodb.ipynb <==
# Read all isotherms from the ISO-DB

import git
import json
import pickle
from tqdm import tqdm
from pathlib import Path

isodb_git_path = Path("/home/daniele/Databases/isodb-library") # Cloned from https://github.com/NIST-ISODB/isodb-library

print("Last commit:", git.Repo(isodb_git_path).head.object.hexsha)

## List of isotherms

We first download lists of all isotherms, materials and adsorbents in the NIST database.

`Isotherms` are dictionaries like:
```
    {
        "filename": "10.1002adem.200500223.isotherm1",
        "DOI": "10.1002/adem.200500223",
        "adsorbent": {
            "hashkey": "NIST-MATDB-991daf7313251e7e607e2bab2da57e33"
        },
        "adsorbates": [
            {
                "InChIKey": "UFHFLCQGNIYNRP-UHFFFAOYSA-N"
            }
        ],
        "category": "",
        "temperature": 77,
        "tabular_data": 0,
        "isotherm_type": ""
    },
```
`Adsorbents` are dictionaries like:
```
    {
        "hashkey": "NIST-MATDB-991daf7313251e7e607e2bab2da57e33",
        "name": "CuBTC",
        "synonyms": [
            "Basolite C300",
            "C300",
            "Cu-BTC",
            "Cu3(BTC)2",
            "HKUST-1",
            "MOF-199"
        ]
    },
```
`Adsorbates` are dictionaries like:
```
    {
        "InChIKey": "CURLTUGMZLYLDI-UHFFFAOYSA-N",
        "name": "Carbon Dioxide",
        "synonyms": [
            "Carbon dioxide",
            "Carbon oxide",
            "CO2"
        ]
    },
```

`Bibliography` are dictionaries like:
```
{
        "DOI": "10.1002/adem.200500223",
        "title": "Improved Hydrogen Storage in the Metal-Organic Framework Cu3(BTC)2",
        "journal": "Advanced Engineering Materials",
        "year": 2006,
        "authors": [
            "Piotr Krawiec",
            "Markus Kramer",
            "Michal Sabo",
            "R. Kunschke",
            "Heidrun Fröde",
            "Stefan Kaskel"
        ],
        "categories": [
            "exp"
        ],
        "adsorbents": [
            {
                "hashkey": "NIST-MATDB-991daf7313251e7e607e2bab2da57e33",
                "name": "CuBTC"
            }
        ],
        "adsorbentMaterial": [
            "CuBTC"
        ],
        "adsorbates": [
            {
                "InChIKey": "IJGRMHOSHXDMSA-UHFFFAOYSA-N",
                "name": "Nitrogen"
            },
            {
                "InChIKey": "UFHFLCQGNIYNRP-UHFFFAOYSA-N",
                "name": "Hydrogen"
            }
        ],
        "adsorbateGas": [
            "Nitrogen",
            "Hydrogen"
        ],
        "temperatures": [
            77,
            87
        ],
        "pressures": [
            0,
            1
        ],
        "isotherms": [
            {
                "filename": "10.1002adem.200500223.isotherm1"
            },
            {
                "filename": "10.1002adem.200500223.isotherm2"
            },
            {
                "filename": "10.1002adem.200500223.isotherm3"
            }
        ]
    }
```

# load files to memory (to copy in other notebook!)
nistdb = {
    'Adsorbents': [],
    'Adsorbates': [],
    'Bibliography': [],
    'Isotherms': [],
}

for key in nistdb:
    if key=='Isotherms':
        for iso_path in tqdm((isodb_git_path / "Library").glob("10*/*")):
            with open(iso_path, "r") as iso_json:
                nistdb[key].append(json.load(iso_json))
    else:
        for json_path in (isodb_git_path / "Library").glob(f"{key}/*"):
            with open(json_path , "r", encoding='utf-8') as json_file:
                nistdb[key].append(json.load(json_file))
    print(f"Loaded {len(nistdb[key])} {key}")

> NOTE: the Adsorbates present in the isotherms are only 280 (not 434), as shown in a later cell

## Cleaning isotherms data
Inspired by 
* Add `total_adsorption`: due to the switch to a multicomponent JSON format, the *total_adsorption* field, which is supposed to show the total amount of all species adsorbed is sometimes left blank. here we iterate and generate this field when it is absent.
* Correct `category=""` when possible: for old data, `category` is left blank. We can correct it only for those isotherm whose biblio is only of one kind. 

# Add total_adsorption
import numpy as np
isos = nistdb['Isotherms']

fixes = 0
for iso in isos:
    if iso['isotherm_data'][0]['total_adsorption'] is None:
        fixes += 1
        for point in iso['isotherm_data']:
            point['total_adsorption'] = float(np.sum([dp['adsorption'] for dp in point['species_data']]))

print(f"Corrections performed in {fixes} isotherms over {len(isos)}.")

# Correct category="": assign 'exp' or 'sim' to all isotherms, if the enclosing paper has only one category!

doi_to_category = { bib['DOI'] : bib['categories'] for bib in  nistdb['Bibliography']}

fixes, known, unknown = 0, 0, 0
for iso in isos:
    if iso['category']=="":
        if iso['DOI'] in doi_to_category and len(doi_to_category[iso['DOI']])==1:
            fixes += 1
            iso['category']=doi_to_category[iso['DOI']][0]
        else:
            unknown += 1
    else:
        known +=1
        

print(f"Corrections performed in {fixes} isotherms, {known} already assigned, and {unknown} left unknown.")

## Save (or load) pickled isotherms if needed

with open("data/nistdb.pickle", 'wb') as f:
    pickle.dump(nistdb, f)

# To copy in other notebooks to load the database

isos = nistdb['Isotherms']
print(f"Loaded {len(isos)} full isotherms.")

# Analyse the database after data-cleaning

print("Total isotherms:", len(nistdb['Isotherms']))
print("Total adsorbents:", len(nistdb['Adsorbents']))
print('Total articles:', len(nistdb['Bibliography']))
adsorbates = set([ ads['name'] for isot in nistdb['Isotherms'] for ads in isot['adsorbates'] ])
print('Total adsorbates:', len(adsorbates))

nistdb['Adsorbents'][100]

# Identity of adsorbents
for match in ["zeolite", "carbon", "mof", "cof"]:
    count = 0
    for adsorbent in nistdb['Adsorbents']:
        if any([ match in x.lower() for x in [adsorbent['name']] + adsorbent['synonyms']]):
            count+=1
    perc = 100 * count/len(nistdb['Adsorbents'])
    print(f"'{match}' matches: {perc:.3f}")


# Some statistics referring to the articles in the Bibliography

bib = nistdb['Bibliography']

count = len([art for art in bib if art['isotherms']])
print('Paper containing at least one isotherm:', count, f"({round(count/len(bib)*100,1)})%")

count = len([art for art in bib if art['isotherms'] and 'exp' in art['categories']])
print('Paper containing at least one isotherm, and HAS experimental:', count, f"({round(count/len(bib)*100,1)})%")

count = len([art for art in bib if art['isotherms'] and art['categories']==['exp']])
print('Paper containing at least one isotherm, and ONLY experimental:', count, f"({round(count/len(bib)*100,1)})%")


# Some statistics on the isoterms
isos = nistdb['Isotherms']
print("Total isotherms:", len(isos))

count = len([isot for isot in isos if isot['category']=='exp'])
print('Experimental:', count, f"({round(count/len(isos)*100,1)})%")

for adsorbate in ['Nitrogen', 'Carbon Dioxide', 'Hydrogen', 'Methane', 'Water']:
    count = len([isot for isot in isos if isot['category']=='exp' and len(isot['adsorbates'])==1 and isot['adsorbates'][0]['name']==adsorbate])
    print(f'Experimental, {adsorbate}:', count, f"({round(count/len(isos)*100,1)})%")

print()
count = len([isot for isot in isos if isot['category']=='exp' and len(isot['adsorbates'])==1 and isot['adsorbates'][0]['name']=='Nitrogen' and isot['temperature']==77])
print(f'Experimental, Nitrogen@77:', count, f"({round(count/len(isos)*100,1)})%")
count = len([isot for isot in isos if isot['category']=='exp' and len(isot['adsorbates'])==1 and isot['adsorbates'][0]['name']=='Argon' and isot['temperature']==87])
print(f'Experimental, Argon@87:', count, f"({round(count/len(isos)*100,1)})%")

# Some statistics on the adrsorbents
count = set()
for isot in nistdb['Isotherms']:
    if isot['category'] == 'exp':
        count.add(isot['adsorbent']['name'])
print("Adsorbents at least one exp isotherm:", len(count), end=" ")
print(f"({100*len(count)/len(nistdb['Adsorbents']):.2f}%)")
experimental_isotherms = count

# Some statistics on the adrsorbents
count = set()
for isot in nistdb['Isotherms']:
    if isot['category'] == 'exp' and (
        (isot['adsorbates'][0]['name']=='Nitrogen' and isot['temperature']==77) or (
         isot['adsorbates'][0]['name']=='Argon'    and isot['temperature']==87)):
        count.add(isot['adsorbent']['name'])
print("Adsorbents with characterization:", len(count), end=" ")
print(f"({100*len(count)/len(nistdb['Adsorbents']):.2f}%)")

# Isotherms with volumtric unit
count=0
for isot in nistdb['Isotherms']:
    if isot['category']=='exp':
        unit = isot['adsorptionUnits']
        if "/" in unit and unit.split("/")[-1] in ['mol', 'volume', 'cm3', 'm3', 'cm2', 'm2', 'unit cell', 'unitcell', 'formula']:
            #print(unit.split("/"))
            count+=1
print("Adsorbents with volumetric (or weird) units:", count, end=" ")
print(f"({100*(count/len(experimental_isotherms)):.2f}%)")

> NOTE: in notebook `06` one can see that for our selected characterization isotherms the volumetric (or weird) units are just <1%

# Find possibly ambiguous names
possible_ambiguous_names = set([])
ambiguous_names = set([])

metals = [
    'Mg', 'Al',
    'Ca', 'Sc', 'Ti', 'V', 'Cr', 'Mn', 'Fe', 'Co', 'Ni', 'Cu', 'Zn', 
    'Zr', 'Ru', 'Rh',
    'Cd', 'In', 'Sn', 'Pb']

for m in metals: 
    for mat in nistdb['Adsorbents']:
        for syn in [mat['name']] + mat['synonyms']:
            if syn.endswith(f"({m})"): # name like: MOF-XX(M)
                possible_ambiguous_names.add(syn.replace(f"({m})",""))
            elif syn.startswith(f"{m}-"): # name like: M-MOF-XX
                possible_ambiguous_names.add(syn.replace(f"{m}-",""))

# Check if there is an ambiguous name which does not specify the metal
for mat in nistdb['Adsorbents']:
    if mat['name'] in possible_ambiguous_names:
        ambiguous_names.add(mat['name'])
        
ambiguous_dict = { k: [] for k in ambiguous_names}

# Link the ambiguous name (where the metal is not specified) to all its variants
for mat in nistdb['Adsorbents']:
    for ambiguous_name in ambiguous_names:
        if any(
            [f"{ambiguous_name}({m})"==syn for m in metals for syn in [mat['name']] + mat['synonyms']]
        ) or any(
            [f"{m}-{ambiguous_name}"==syn for m in metals for syn in [mat['name']] + mat['synonyms']]):
            ambiguous_dict[ambiguous_name].append(mat['name'])

for k in sorted(ambiguous_dict):
    print(f"{k:20s} >>> ", ambiguous_dict[k])
# NOTE: the selection of names to exclude will be manual (see Supporting Info of the draft).
