==> 02_scrape_csd.ipynb <==
# Scrape CSD and curate the synonyms
NOTE: you need a CSD licence to install and use the CSD API

from pathlib import Path
from tqdm import tqdm
import pandas as pd

import ccdc
import ccdc.io
import ccdc.search
print("ccdc.__version__", ccdc.__version__)
print("ccdc.io.csd_directory()", ccdc.io.csd_directory())
print("ccdc.io.csd_version()", ccdc.io.csd_version())

subset_mofs_all = Path(ccdc.io.csd_directory()) / "subsets" / "MOF_subset.gcd"
print(f"Number of MOFs: {len(ccdc.io.EntryReader(str(subset_mofs_all)))}")

n_entries = len(ccdc.io.EntryReader(str(subset_mofs_all)))
rows = [{} for _ in range(n_entries)] # Pre allocate for speed
for i, entry in enumerate(tqdm(ccdc.io.EntryReader(str(subset_mofs_all)))): # Takes 2 minutes
    for col in ['identifier', 'ccdc_number','formula', 'has_disorder']:
        rows[i][col] = getattr(entry, col)
    rows[i]['publication_year'] = entry.publication.year
    doi = entry.publication.doi
    if doi: # can be None
        doi = doi.lower() # Lower caption to improve match in a later step
    rows[i]['publication_doi'] = doi 
    synonyms = list(entry.synonyms)
    rows[i]['synonyms_orig'] = synonyms
df = pd.DataFrame(rows) 

df.to_csv("data/step-00.csv", index=False) 
# step-00.csv contains all CSD entries of MOF subset: id, formula, has_disorder, pub year, doi, synonyms_orig

df.head(3)

df = pd.read_csv("data/step-00.csv")
print("CSD entries having synonyms:", len(df[df['synonyms_orig']!="[]"]))

# Correct the names from "synonyms_orig", and create a new column "synonyms": exclude misleading ones, and solvents/adsorbate.
# NOTE: the reason why I'm doing it here and not later in the name match, is to exclude those names that are obviously problematic,
#       and I want to remove completely, to make later checks easier.

ADSORBATES_LIST = [
    "carbon", # carbon dioxide
    "deuteromethane",
    "methane",
    "ethane",
    "ethene",
    "dinitrogen",
    "acetylene",
    "acetylene",
    "acetaldehyde",
    "ethylamine"  
]

df = pd.read_csv("data/step-00.csv")
df['synonyms'] = None # create new column for .at
for i, synonyms in enumerate(tqdm(df['synonyms_orig'])):
    synonyms_valid = [] 
    for name in eval(synonyms): #there may be more than one synonym, but it is not very frequent
        # Exclude some fake names:
        if any([
            name.startswith('catena'),
            name.startswith('Teaching Subset'),
            name.startswith('DrugBank')
            #len(name)>20
        ]):
            continue
            
        if name in ["","MOF-1", "MOF-2", 1, 2, "1", "2"]:
            continue
            
        name_split = name.split(" ")
        if len(name_split)>1 and name_split[1] in ADSORBATES_LIST: # remove adsorbent from name
            name = name_split[0] 
            
        synonyms_valid.append(name)
        
    df.at[i,'synonyms'] = synonyms_valid

df.to_csv("data/step-01.csv", index=False) 
# step-01.csv: add "synonyms" column with corrected CSD syns

df = pd.read_csv("data/step-01.csv")
print("CSD entries having valid synonyms:", len(df[df['synonyms']!="[]"]))

df = pd.read_csv("data/step-01.csv")
df[df['synonyms']!="[]"].head(3)

# Inspect names that have a space: one can see that many more can be fixed but it is hard to find general rules
df = pd.read_csv("data/step-01.csv")
for i in df.index:
    synonyms = eval(df.at[i,'synonyms'])
    if len(synonyms)>0:
        first = synonyms[0]
        if len(first.split(" "))>1:
            print(first)

# Check papers with a lot of entries
doi_note = {
    "10.1002/cssc.201601752": "in-situ study of M-MOF-74",
}

for doi, note in doi_note.items():
    df_doi = df[df['publication_doi']==doi]
    print(doi, f"({note})", f"Number of CIFs: {len(df_doi)}", 'Synonyms:', *list(set(df_doi['synonyms'])))

# Keep only 3 MOFs for the same DOI with the same synonyms, to avoid having paper with hundreds of structures
df = pd.read_csv("data/step-01.csv")
df['note'] = "-" # Assign a string for all the notes, to make the later filtering easier
df = df.sort_values(by=['publication_doi','synonyms','identifier'])
df = df.reset_index() # keeping the old as column "index"

max_same = 3
count_same=0
for i in tqdm(df.index[1:]): # skipping the first because it is comparing with the previous (no privious for the first!)
    if df.at[i,'synonyms']!='[]':
        if df.at[i,'publication_doi']==df.at[i-1,'publication_doi'] and df.at[i,'synonyms']==df.at[i-1,'synonyms']:
            count_same+=1
        else:
            count_same=0
        if count_same>=3:
            df.at[i, 'note'] = f"Excluding more than {max_same} same MOFs from the same DOI"

df = df.sort_values(by='index')
df = df.drop(columns='index') # remove the dummy column
df.to_csv("data/step-02.csv", index=False)
# step-02.csv: contains a "note" for excluding structures if >3 from the same paper

df = pd.read_csv("data/step-02.csv")
len_after_exclusions = len(df[(df['synonyms']!="[]") & (df['note']=='-')])
print("CSD entries having synonyms:", len_after_exclusions, f"(including only {max_same} same materials per paper)")

df = pd.read_csv("data/step-02.csv")
df[df['synonyms']!="[]"].head(3)
