==> 04_match_by_doi.ipynb <==
# Match by DOI

import pandas as pd
import pickle 
from tqdm import tqdm
import json

with open("data/nistdb.pickle", 'rb') as f:
    nistdb = pickle.load(f)
    
with open("data/csd_542_mofchecker.json", "r") as file:
    mof_checker = json.load(file)

mof_checker['FIQCEN']

# Add a column with nist_mat matched by doi

df = pd.read_csv("data/step-04.csv")

for pub in nistdb['Bibliography']:
    pub['DOI'] = pub['DOI'].lower() # NOTE: I already lowered all DOIs from CSD

for i, doi_csd in enumerate(tqdm(df['publication_doi'])): # takes 1 minute
    if not doi_csd: # Missing DOI
        continue
    for pub in nistdb['Bibliography']:
        doi_nist = pub['DOI']
        if doi_csd==doi_nist:
            df.at[i,"nist_mat_by_doi"] = str(pub['adsorbentMaterial']) # NOTE 1: adsorbentMaterial will be always the main name for NIST, not a synonym (e.g., no HKUST-1, only CuBTC)
            break                                                      # NOTE 2: pub['adsorbentMaterial'] is a list that may contain more entries if the paper contains more materials!
df.to_csv("data/step-05.csv", index=False) 
# step-05.csv: add column nist_mat_by_doi with possible matches but not performing the match yet 

df = pd.read_csv("data/step-05.csv")
df[~pd.isna(df.nist_mat_by_doi)].head(3)

# Showing (but not doing anything) possible names that need correction: 
#   they match DOI and have CSD synonyms, but their synonyms are not matching
df = pd.read_csv("data/step-05.csv")
df = df[(df['note']=='-') & ~pd.isnull(df['nist_mat_by_doi']) & (df['synonyms']!='[]')]
df[['identifier','publication_doi','synonyms','nist_mat_by_doi']].reset_index()

# Flag the ones with multiple NIST-ISODB adsorbents, but if unique add them to the nist_mat
df = pd.read_csv("data/step-05.csv")

for row in tqdm(df.itertuples(), total=df.shape[0]):
    if row.note.startswith("Matched by synonym"):
        continue
    elif row.note=="Exclude DOI already matched by synonym":
        continue        
    elif row.note=="Excluding more than 3 same MOFs from the same DOI":
        continue
    elif pd.isnull(row.nist_mat_by_doi):
        continue
        
    else:
        synonyms = eval(row.nist_mat_by_doi)
        if len(synonyms)==1: # Only one nist-isodb
            df.at[row.Index,"nist_mat"] = synonyms[0]
            if len(df[df.publication_doi==row.publication_doi])==1: # check if there are multiple CSD from the same DOI
                df.at[row.Index,"note"] = "One-to-one match by DOI"
            else:
                df.at[row.Index,"note"] = "OneNIST-to-manyCSD matches by DOI (to check)"
                
        # Third choice, if there is >one nist_adsorbent_by_doi, write them all and mark that there is need for revision
        else:
            df.at[row.Index,"note"] = "VariousNIST-to-paper matched by DOI"

df.to_csv("data/step-06.csv", index=False) 
# step-06.csv: add nist_mat for unique match, and "note": still need to check for multiple CSD

df = pd.read_csv("data/step-06.csv")
df[~pd.isna(df.nist_mat_by_doi)].head(3)

# Show statistics
df = pd.read_csv("data/step-06.csv")
df = df.sort_values('note', ascending=True) # When I keep=first, these are sorted by note

print('Statistics by CSD entries')
for note in df['note'].unique():
    count = len(df[df.note==note])
    print(count, "\t",note)
    
print('\nStatistics by Paper')
df1 = df[~pd.isnull(df['publication_doi'])]
df1 = df1[~df1.duplicated(subset='publication_doi', keep='first')]

for note in df['note'].unique():
    count = len(df1[df1.note==note])
    print(count, "\t",note)
    
print('\nStatistics by NIST matches')
df = df[~pd.isnull(df['nist_mat'])]
df = df[~df.duplicated(subset='nist_mat', keep='first')]

tot = 0
for note in df['note'].unique():
    count = len(df[df.note==note])
    tot+=count
    print(count, "\t",note)
print(tot, "\t","TOTAL")
    
print("\nNOTE: You can find some little mismatch in the NIST stats because df.duplicates() may remove rows from different categories,")
print("       but I used df.sort_values and ascending Notes to avoid it.")

# Add structure SGH (Structure Graph Hash) for AllSolventRemoved structures
df = pd.read_csv("data/step-06.csv")

missing = 0
for i, csd_id in enumerate(tqdm(df['identifier'])):
    try:
        graph = mof_checker[csd_id]['asr']['graph_hash']
    except:
        missing+=1
        graph = "MISSING_" + csd_id   
    df.at[i, 'sgh_asr'] = graph

print('Missing SGH:', missing)
df.to_csv("data/step-07.csv", index=False) 
# step-07.csv: add SGH

df = pd.read_csv("data/step-07.csv")
df.head(3)

# Identify if CSD structures linked to the same papers are actually the same once desolvated (ASR)

PRINT = True    

df = pd.read_csv("data/step-07.csv")
df_to_check = df[df['note'] == "OneNIST-to-manyCSD matches by DOI (to check)"]

# Loop over the papers with same DOI
all_identifiers = []
for i, doi in enumerate(set(df_to_check['publication_doi'])):
    df_same_doi = df[df['publication_doi'] == doi]
    nist_mat = df_same_doi['nist_mat'].values[0]
    all_identifiers+= list(df_same_doi['identifier'].values)
    if PRINT: print(i, nist_mat, doi)
    graphs = df_same_doi['sgh_asr'].values
    if all([x==graphs[0] for x in graphs]):
        if PRINT: print(' >>> ALL THE SAME!', *df_same_doi['identifier'].values)
        for index in df_same_doi.index:
            df.at[index, "note"] = "OneNIST-to-manyCSD matches by DOI >> MATCH, same SGH"
    else:
        if PRINT: print(' >>> SOME DIFFERENT, REMOVE!', *df_same_doi['identifier'].values)
        
        for index in df_same_doi.index:
            df.at[index, 'nist_mat'] = None
            df.at[index, "note"] = "OneNIST-to-manyCSD matches by DOI >> EXCLUDE, different SGH"

df.to_csv("data/step-08.csv", index=False) 
# data/step-08.csv: filter entries with multiple 

df = pd.read_csv("data/step-08.csv")
df[df.note.str.startswith("OneNIST-to-manyCSD")].head(3)

# Shows statistics
df = pd.read_csv("data/step-08.csv")
df = df.sort_values('note', ascending=True) # When I keep=first, these are sorted by note

print('Statistics by CSD entries')
for note in set(df['note']):
    count = len(df[df.note==note])
    print(count, "\t",note)
    
tot = df[df.note.isin([
    'Matched by synonym', 
    'Matched by synonym (corrected)',
    'One-to-one match by DOI',
    'OneNIST-to-manyCSD matches by DOI >> MATCH, same SGH',
])]  

print("\n Total CSD matched:", len(tot))
    
print()
print('Statistics by NIST matches')
df = df[~pd.isnull(df['nist_mat'])]
df = df[~df.duplicated(subset='nist_mat', keep='first')]
tot=0
for note in set(df['note']):
    count = len(df[df.note==note])
    tot+=count
    print(count, "\t",note)

print("\n Total NIST matched:", tot)
