==> 05_check_structures.ipynb <==
# Check crystal structures and remove the ones with overlapping atoms

import pandas as pd
import pickle 
from tqdm import tqdm
import json
import yaml

MANUAL_EXCLUSIONS = list(pd.read_csv("data/manual_exclusions.csv")['identifier'])

with open("data/csd_542_mofchecker.json", "r") as file:
    mof_checker = json.load(file)
    
# fix
for mof, mof_dict in mof_checker.items():
    for structure in mof_dict:
        mof_checker[mof][structure]['has_atomic_overlaps'] = not mof_checker[mof][structure]['has_atomic_overlaps']
        mof_checker[mof][structure]['has_not_hydrogen'] = not mof_checker[mof][structure]['has_hydrogen']
        del mof_checker[mof][structure]['has_hydrogen']

mof_checker['FIQCEN']['asr']

df = pd.read_csv("data/step-08.csv").dropna(subset=['nist_mat'])
df = df[~df.identifier.isin(MANUAL_EXCLUSIONS)]                        

mat_sel = list(set(df['nist_mat']))
mat_sel.sort()
mat_dict = {}
csd_refcodes_mem = []
for nist_name in mat_sel:
    csd_refcodes = list(df[df['nist_mat']==nist_name]['identifier'].values)
    csd_refcodes_mem += csd_refcodes
    mat_dict[nist_name] = {}
    for csd_refcode in csd_refcodes:       
        mat_dict[nist_name][csd_refcode] = df[df['identifier']==csd_refcode]['sgh_asr'].values[0]
        
with open("data/step-09.yml", "w") as f:
    yaml.dump(mat_dict, f)
# step-09.yml: { nist_adsorbent1: [csd1: SGH, csd2: SGH, ...], ... }
    
print("... after manual exclusions ...")    
print(f'All identified NIST-ISODB adsorbents: {len(mat_dict)}')
print(f'All identified CSD entries: {len(csd_refcodes_mem)}')

with open("data/step-09.yml", "r") as file:
    mat_dict = yaml.safe_load(file)
mat_dict['CuBTC']

# check the ones that have same nist_mat but different graph
# - used to verify and create manual_exclusions.csv
# - some are left because by visual inspection they are the same

with open("data/step-09.yml", "r") as file:
    mat_dict = yaml.safe_load(file)

count_unique, count_multiple, count_tocheck = 0, 0, 0
to_check = []
for mat_name, d in mat_dict.items():
    if len(d)==0:
           print('ERROR', mat_name)
    elif len(d)==1:
        count_unique+=1
    #elif len(d)<30:
    else:
        count_multiple+=1
        graph_values = list(d.values())
        if not all( [ x==graph_values[0] for x in graph_values]):
            print(mat_name)
            print(d)
            to_check+=list(d.keys())
            count_tocheck+=1
            
print()
print('NOTE: these left are ok, possibly primitive/conventional cells ')
print()
print("Unique NIST - CSD matches:", count_unique)
print("NIST - multiple CSD matches:", count_multiple)
print("NIST - CSD matches, total:", count_unique+count_multiple)
print("NIST - CSD matches to check because of different graphs:", count_tocheck)

# Inspection

with open("data/step-09.yml", "r") as file:
    mat_dict = yaml.safe_load(file)
    
count = {
    'has_not_hydrogen': [],
    'has_atomic_overlaps': [],
    'has_lone_molecule': [],    
}

missing = []
for csd in csd_refcodes_mem:
    if csd not in mof_checker:
        missing.append(csd)
        continue
    for k in count:
        if mof_checker[csd]['asr'][k]:
            count[k].append(csd)

count['missing'] = missing
count['total'] = csd_refcodes_mem

for k in count:
    print(k, len(count[k]))

with open("data/step-09.yml", "r") as file:
    mat_dict = yaml.safe_load(file)
    
mat_dict_new = {}
for nist_mat, nist_dict in mat_dict.items():
    nist_dict_new = {}
    for csd_mat in nist_dict:  
        if not csd_mat in count['has_atomic_overlaps'] + count['has_lone_molecule']:
            nist_dict_new[csd_mat] = nist_dict[csd_mat]
    if len(nist_dict_new)>0:
        mat_dict_new[nist_mat] = nist_dict_new
        
with open("data/step-10.yaml", "w") as f: # Same as -40 but without CSD entries with problems (overlaps)
    yaml.dump(mat_dict_new, f)  
    
print("NIST adsorbents left after cleaning:", len(mat_dict_new) ) 
print("CSD entries adsorbents left after cleaning:", sum([ len(v) for v in mat_dict_new.values()]) )
