==> 07_explorer.ipynb <==
# Explorer

import pickle
import yaml

with open("data/nistdb.pickle", 'rb') as f:
    nistdb = pickle.load(f)
    
with open("data/step-12-geometric.yml") as f:
    pv = yaml.safe_load(f)
    
pv['CuBTC']

# Get all the isotherms from valid DOIs
# i.e., containing characterization isotherms for which the MPV/CPV ratio is in a certain threshold

MOF = 'CuBTC'
MIN_THR_RATIO  = 0.75
MAX_THR_RATIO  = 1.10

max_pv_from_structure = max(pv[MOF]['structures'].values())
valid_dois = [ isot.split(".iso")[0] 
              for isot, mpv in pv[MOF]['isotherms'].items()
              if MIN_THR_RATIO<mpv/max_pv_from_structure<MAX_THR_RATIO ] 

for isot in nistdb['Isotherms']:
    if isot['adsorbent']['name']==MOF and isot['DOI'].lower().replace("/","") in valid_dois:
        print(MOF, isot['adsorbates'][0]['name'], isot['temperature'], isot['filename'])
